+++ source/select.js
@@ -1329,11 +1329,6 @@ const styles = styler
 		user-select         : none
 
 	list
-		position        : absolute
-		z-index         : 1
-		margin-top      : 0
-		margin-bottom   : 0
-		padding         : 0
 		list-style-type : none
 		overflow-x      : hidden